BinaryBeast/Project/AnyERP/requirements.txt
django-filter==2.4.0
django==3.0.8
openpyxl~=3.0.4
convertapi~=1.4.0
djangorestframework==3.11.2
django-debug-toolbar~=3.2